LabeljsMetric: Apply label information to jstable metric object using label data

Apply label information to metric object of jstable using label data

LabeljsMetric(obj.metric, ref)

Arguments

obj.metric
metric of lmer.display, coxme.display

ref
Label data made by mk.lev function

Value

metric of lmer.display, coxme.display with label information

Details

DETAILS

Examples

library(coxme)

#> Loading required package: bdsmatrix

#> 
#> Attaching package: ‘bdsmatrix’

#> The following object is masked from ‘package:base’:
#> 
#>     backsolve

fit <- coxme(Surv(time, status) ~ sex + ph.ecog + ph.karno + (1 | inst) + (1 | sex), lung)
fit.table <- coxme.display(fit)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

lung.label <- mk.lev(lung)
LabeljsTable(fit.table$table, ref = lung.label)

#>          crude HR(95%CI)    crude P value adj. HR(95%CI)     adj. P value
#> sex      "0.59 (0.42,0.82)" "0.002"       "0.56 (0.4,0.79)"  "< 0.001"   
#> ph.ecog  "1.66 (1.32,2.09)" "< 0.001"     "1.96 (1.38,2.8)"  "< 0.001"   
#> ph.karno "0.98 (0.97,1)"    "0.005"       "1.01 (0.99,1.03)" "0.241"     

LabeljsRanef(fit.table$ranef, ref = lung.label)

#>                 [,1] [,2] [,3] [,4]
#> Random effect     NA   NA   NA   NA
#> inst(Intercept) 0.02   NA   NA   NA
#> sex(Intercept)  0.00   NA   NA   NA

LabeljsMetric(fit.table$metric, ref = lung.label)

#>                     [,1] [,2] [,3] [,4]
#> <NA>                  NA   NA   NA   NA
#> No. of group(inst)    18   NA   NA   NA
#> No. of group(sex)      2   NA   NA   NA
#> No. of observations  225   NA   NA   NA
#> No. of events        162   NA   NA   NA