LabeljsRanef: Apply label information to jstable random effect object using label data

Apply label information to ranef object of jstable using label data

LabeljsRanef(obj.ranef, ref)

Arguments

obj.ranef
ranef of lmer.display, coxme.display, cox2.display

ref
Label data made by mk.lev function

Value

ranef of lmer.display, coxme.display, cox2.display with label information

Details

DETAILS

Examples

# NOT RUN {
if(interactive()){
 library(coxme)
 fit <- <- coxme(Surv(time, status) ~ sex + ph.ecog + ph.karno + (1|inst) +(1|sex), lung)
 fit.table = coxme.display(fit)
 lung.label = mk.lev(lung)
 LabeljsTable(fit.table$table, ref = lung.label)
 LabeljsRanef(fit.table$ranef, ref = lung.label)
 }
# }