LabeljsTable: Apply label information to jstable object using label data

Apply label information to table of geeglm.display, lmer.display, coxme.display using label data

Usage

LabeljsTable(obj.table, ref)

Arguments

obj.table

table of geeglm.display, lmer.display, coxme.display

ref

Label data made by mk.lev function

Value

table of geeglm.display, lmer.display, coxme.display with label information

Details

DETAILS

Examples

library(coxme)
fit <- coxme(Surv(time, status) ~ sex + ph.ecog + ph.karno + (1 | inst) + (1 | sex), lung)
fit.table <- coxme.display(fit)
#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.
#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.
#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.
lung.label <- mk.lev(lung)
LabeljsTable(fit.table$table, ref = lung.label)
#>          crude HR(95%CI)    crude P value adj. HR(95%CI)     adj. P value
#> sex      "0.59 (0.42,0.82)" "0.002"       "0.56 (0.4,0.79)"  "< 0.001"   
#> ph.ecog  "1.66 (1.32,2.09)" "< 0.001"     "1.96 (1.38,2.8)"  "< 0.001"   
#> ph.karno "0.98 (0.97,1)"    "0.005"       "1.01 (0.99,1.03)" "0.241"