svyCreateTableOneJS: Modified CreateTableOne function in tableone package

Combine svyCreateTableOne & print function in tableone package

Usage

svyCreateTableOneJS(
  vars,
  strata = NULL,
  strata2 = NULL,
  data,
  factorVars = NULL,
  includeNA = F,
  test = T,
  showAllLevels = T,
  printToggle = F,
  quote = F,
  smd = F,
  Labels = F,
  nonnormal = NULL,
  catDigits = 1,
  contDigits = 2,
  pDigits = 3,
  labeldata = NULL,
  psub = T,
  minMax = F,
  showpm = T,
  addOverall = F,
  pairwise = F,
  pairwise.showtest = F,
  n_original = T
)

Arguments

vars

Variables to be summarized given as a character vector. Factors are handled as categorical variables, whereas numeric variables are handled as continuous variables. If empty, all variables in the data frame specified in the data argument are used.

strata

Stratifying grouping variable name(s) given as a character vector. If omitted, the overall results are returned.

strata2

Stratifying 2nd grouping variable name(s) given as a character vector. If omitted, the 1 group results are returned.

data

A data frame in which these variables exist. All variables (both vars and strata) must be in this data frame.

factorVars

Numerically coded variables that should be handled as categorical variables given as a character vector. Do not include factors, unless you need to relevel them by removing empty levels. If omitted, only factors are considered categorical variables. The variables specified here must also be specified in the vars argument.

includeNA

If TRUE, NA is handled as a regular factor level rather than missing. NA is shown as the last factor level in the table. Only effective for categorical variables., Default: F

test

If TRUE, as in the default and there are more than two groups, groupwise comparisons are performed, Default: T

showAllLevels

Whether to show all levels. FALSE by default, i.e., for 2-level categorical variables, only the higher level is shown to avoid redundant information., Default: T

printToggle

Whether to print the output. If FALSE, no output is created, and a matrix is invisibly returned., Default: F

quote

Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily, Default: F

smd

If TRUE, as in the default and there are more than two groups, standardized mean differences for all pairwise comparisons are calculated, Default: F

Labels

Use Label, Default: F

nonnormal

A character vector to specify the variables for which the p-values should be those of nonparametric tests. By default all p-values are from normal assumption-based tests (oneway.test)., Default: NULL

catDigits

Number of digits to print for proportions., Default: 1

contDigits

Number of digits to print for continuous variables. Default 2.

pDigits

Number of digits to print for p-values (also used for standardized mean differences), Default: 3

labeldata

labeldata to use, Default: NULL

psub

show sub-group p-values, Default: F

minMax

Whether to use [min,max] instead of [p25,p75] for nonnormal variables. The default is FALSE.

showpm

Logical, show normal distributed continuous variables as Mean ± SD. Default: T

addOverall

(optional, only used if strata are supplied) Adds an overall column to the table. Smd and p-value calculations are performed using only the stratifed clolumns. Default: F

pairwise

(optional, only used if strata are supplied) When there are three or more strata, it displays the p-values for pairwise comparisons. Default: F

pairwise.showtest

(optional, only used if strata are supplied) When using pairwise comparison, it displays the test used to calculate p-values for pairwise comparisons. Default: F

n_original

Replace the number of weighted n with the n in the original data. Default: T

Value

A matrix object containing what you see is also invisibly returned. This can be assinged a name and exported via write.csv.

Details

DETAILS

Examples

library(survey)
data(nhanes)
nhanes$SDMVPSU <- as.factor(nhanes$SDMVPSU)
nhanesSvy <- svydesign(
  ids = ~SDMVPSU, strata = ~SDMVSTRA, weights = ~WTMEC2YR,
  nest = TRUE, data = nhanes
)
svyCreateTableOneJS(
  vars = c("HI_CHOL", "race", "agecat", "RIAGENDR"),
  strata = "RIAGENDR", data = nhanesSvy,
  factorVars = c("HI_CHOL", "race", "RIAGENDR")
)
#> $table
#>              level      1                 2                 p        test sig 
#> n            ""         "4247.0"          "4344.0"          ""       ""   NA  
#> HI_CHOL (%)  "0"        "3526.0 ( 89.9) " "3533.0 ( 87.7) " " 0.010" ""   "**"
#>              "1"        " 363.0 ( 10.1) " " 424.0 ( 12.3) " ""       ""   NA  
#> race (%)     "1"        "1339.0 ( 15.8) " "1378.0 ( 14.3) " " 0.042" ""   "**"
#>              "2"        "1863.0 ( 66.2) " "1880.0 ( 65.3) " ""       ""   NA  
#>              "3"        " 798.0 ( 11.1) " " 825.0 ( 12.7) " ""       ""   NA  
#>              "4"        " 247.0 (  6.8) " " 261.0 (  7.7) " ""       ""   NA  
#> agecat (%)   "(0,19]"   "1318.0 ( 21.7) " "1214.0 ( 19.9) " " 0.001" ""   "**"
#>              "(19,39]"  " 953.0 ( 30.0) " "1080.0 ( 28.7) " ""       ""   NA  
#>              "(39,59]"  " 995.0 ( 30.4) " "1026.0 ( 30.2) " ""       ""   NA  
#>              "(59,Inf]" " 981.0 ( 17.9) " "1024.0 ( 21.2) " ""       ""   NA  
#> RIAGENDR (%) "1"        "4247.0 (100.0) " "   0.0 (  0.0) " "<0.001" ""   "**"
#>              "2"        "   0.0 (  0.0) " "4344.0 (100.0) " ""       ""   NA  
#> 
#> $caption
#> [1] "Stratified by RIAGENDR- weighted data"
#>